The whole thing
I understand this is not really what ember was made to do. But it helped
me understand the setup and all I was really looking for here was to get
ember to work. That being said, this was pretty frustrating to build
something so simple.
This project helped me understand a bit more of the naming convention
along with views and controllers. The Router, models, and how to really
implement multiple templates is what's really giving me a headache